Add BPC fetching
